removed other codes
